feat: make World Cup picks inline on the group Picks tab
A world_cup_2026 group's Picks tab previously only linked out to the
standalone /world-cup page, so making picks meant leaving the group. The
pick-making surface (stage-grouped match list + sticky submit bar +
multi-group save dropdown) is now extracted from WorldCupPicksPage into
WorldCupPicksTab and embedded directly in the group detail page's Picks tab.

The sticky save bar lives inside the tab component, so it appears when the
user enters the Picks tab and disappears when they switch to another tab.
WorldCupPicksPage stays as a thin shell around the same component, keeping
existing /world-cup deep links working. NFL pools are unchanged.

Tests: the tab's full interaction coverage (stage fetch, pick toggling,
hydrate from saved picks, submit + fan-out, error/empty states) moves to
WorldCupPicksTab.test.tsx; the page test pins the shell behavior; new
GroupDetailsPage tests assert the embedded surface and that the save bar
mounts/unmounts with the tab; the Playwright e2e now drives the whole
in-group flow (enter tab -> bar appears -> pick -> submit hits the group's
world-cup endpoint -> leave tab -> bar disappears).
