+++ frontend/src/pages/WorldCupPicksPage.tsx
@@ -1,59 +1,13 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import Button from '../designsystem/components/Button';
-import InlineToast from '../designsystem/components/InlineToast';
-import type { ToastVariant } from '../designsystem/components/InlineToast/InlineToast';
-import MatchPickRow from '../designsystem/components/MatchPickRow';
-import {
-  WORLD_CUP_STAGES,
-  type MatchPick,
-  type MatchPickResult,
-  type WorldCupMatch,
-  type WorldCupStage,
-} from '../lib/types';
-import {
-  getStageMatches,
-  submitWorldCupPicks,
-  getMyWorldCupPicks,
-} from '../lib/worldCupService.js';
-import { getMyGroups } from '../lib/groupsService.js';
-import SaveTargetsDropdown, { type SaveTarget } from '../components/SaveTargetsDropdown';
+import WorldCupPicksTab from './GroupDetails/WorldCupPicksTab';
 
-// World Cup sibling of GamesPage. A world_cup_2026 pool renders the WHOLE
-// tournament as one stage-grouped match list — there is NO week selector. The
-// group identifier comes from the `group` query param (mirrors
-// GroupDetailsPage), and a member's picks are flat (home/away/draw) with no
-// confidence — World Cup scoring is flat-per-match (see
-// world-cup-picks-rules.md).
-//
-// We fetch every stage in parallel and group the flattened matches back by
-// stage for rendering, iterating WORLD_CUP_STAGES so sections appear in
-// tournament order regardless of fetch resolution order.
-
-// Section header labels per stage.
-const STAGE_LABEL: Record<WorldCupStage, string> = {
-  group: 'Group Stage',
-  r32: 'Round of 32',
-  r16: 'Round of 16',
-  qf: 'Quarterfinals',
-  sf: 'Semifinals',
-  third: 'Third Place',
-  final: 'Final',
-};
-
-type DraftMap = Record<number, MatchPickResult>;
-
-interface PageState {
-  loading: boolean;
-  error: string;
-  matches: WorldCupMatch[];
-}
-
-interface ToastState {
-  open: boolean;
-  message: string;
-  variant: ToastVariant;
-}
+// World Cup sibling of GamesPage. The pick-making surface itself (stage-grouped
+// match list + sticky submit bar) lives in WorldCupPicksTab so a
+// world_cup_2026 group's Picks tab embeds the same experience inline; this
+// route is the standalone shell around it, kept so existing /world-cup deep
+// links keep working. The group identifier comes from the `group` query param
+// (mirrors GroupDetailsPage).
 
 // Shared not-found UI, mirroring GroupDetailsPage: shown when the `group` query
 // param is absent so the user always lands on a single recoverable error with a
@@ -78,184 +32,6 @@ export default function WorldCupPicksPage() {
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
   const identifier = searchParams.get('group');
-  // The group identifier doubles as the submit target — submitWorldCupPicks
-  // keys off the same value GroupDetailsPage resolves from `?group=`.
-  const groupId = identifier;
-
-  const [pageState, setPageState] = useState<PageState>({
-    loading: false,
-    error: '',
-    matches: [],
-  });
-  const [draft, setDraft] = useState<DraftMap>({});
-  const [submitting, setSubmitting] = useState(false);
-  const [toast, setToast] = useState<ToastState>({ open: false, message: '', variant: 'info' });
-
-  const fetchMatches = useCallback(async () => {
-    setPageState((s) => ({ ...s, loading: true, error: '' }));
-    try {
-      // Fetch every stage in parallel; flatten into one list and group by stage
-      // at render time so the sections follow WORLD_CUP_STAGES order.
-      const responses = await Promise.all(WORLD_CUP_STAGES.map((stage) => getStageMatches(stage)));
-      const matches = responses.flatMap((r) => (Array.isArray(r?.games) ? r.games : []));
-      setPageState({ loading: false, error: '', matches });
-    } catch (err) {
-      setPageState((s) => ({
-        ...s,
-        loading: false,
-        error: `Failed to fetch matches: ${err instanceof Error ? err.message : 'unknown error'}`,
-      }));
-    }
-  }, []);
-
-  useEffect(() => {
-    if (!identifier) return;
-    fetchMatches();
-  }, [identifier, fetchMatches]);
-
-  // Hydrate the draft from the user's already-saved picks on this group so
-  // a refresh doesn't blank them out. Runs independently of the stage fetch
-  // so a transient picks-endpoint failure can't block the matches from
-  // rendering — the draft just stays empty in that case.
-  useEffect(() => {
-    if (!groupId) {
-      setDraft({});
-      return;
-    }
-    let cancelled = false;
-    getMyWorldCupPicks(groupId)
-      .then((resp) => {
-        if (cancelled) return;
-        const next: DraftMap = {};
-        for (const p of resp.picks ?? []) {
-          if (p && p.gameId != null && p.pickedResult) next[p.gameId] = p.pickedResult;
-        }
-        setDraft(next);
-      })
-      .catch(() => {
-        // Best-effort — don't surface a toast for the silent hydrate; the user
-        // can still re-pick and submit. Leaving draft as whatever it currently is.
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [groupId]);
-
-  function pickResult(matchId: number, result: MatchPickResult) {
-    setDraft((prev) => {
-      const next = { ...prev };
-      // Re-picking the selected outcome clears it; otherwise set the new outcome.
-      if (next[matchId] === result) delete next[matchId];
-      else next[matchId] = result;
-      return next;
-    });
-  }
-
-  // Group the flattened matches by stage, preserving tournament order. Only
-  // stages with at least one match produce a section.
-  const stageGroups = useMemo(
-    () =>
-      WORLD_CUP_STAGES.map((stage) => ({
-        stage,
-        matches: pageState.matches.filter((m) => m.stage === stage),
-      })).filter((g) => g.matches.length > 0),
-    [pageState.matches],
-  );
-
-  const picks: MatchPick[] = useMemo(
-    () =>
-      Object.entries(draft).map(([gameId, pickedResult]) => ({
-        gameId: Number(gameId),
-        pickedResult,
-      })),
-    [draft],
-  );
-
-  const canSubmit = !!groupId && picks.length > 0 && !submitting;
-
-  // Multi-select "save to which World Cup groups?" — restores the heritage
-  // Svelte PR #37 dropdown UX. We eagerly load the user's groups on mount so
-  // the dropdown shows a real count without waiting for the first click.
-  const [myGroups, setMyGroups] = useState<SaveTarget[]>([]);
-  const [groupsLoading, setGroupsLoading] = useState(false);
-  const [selectedTargets, setSelectedTargets] = useState<Set<string>>(
-    () => new Set(groupId ? [groupId] : []),
-  );
-
-  useEffect(() => {
-    let cancelled = false;
-    if (!groupId) return;
-    setGroupsLoading(true);
-    getMyGroups()
-      .then((groups) => {
-        if (cancelled) return;
-        const wc: SaveTarget[] = groups
-          .filter((g) => g.poolType === 'world_cup_2026')
-          .map((g) => ({ identifier: g.identifier, name: g.name }));
-        // Ensure source group is in the list even if the backend didn't return
-        // it (offline-cached state, race with create-group, etc.).
-        if (!wc.some((g) => g.identifier === groupId)) {
-          wc.unshift({ identifier: groupId, name: groupId });
-        }
-        setMyGroups(wc);
-      })
-      .catch(() => {
-        // Fall back to a single-group dropdown — the source group always works.
-        if (!cancelled) setMyGroups([{ identifier: groupId, name: groupId }]);
-      })
-      .finally(() => {
-        if (!cancelled) setGroupsLoading(false);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [groupId]);
-
-  async function submit() {
-    if (!groupId || picks.length === 0) return;
-    setSubmitting(true);
-    try {
-      const targets = Array.from(selectedTargets);
-      if (!targets.includes(groupId)) targets.push(groupId);
-      if (targets.length === 1) {
-        await submitWorldCupPicks(targets[0], picks);
-        setToast({ open: true, message: 'Picks saved', variant: 'success' });
-        return;
-      }
-      const results = await Promise.allSettled(
-        targets.map((id) => submitWorldCupPicks(id, picks)),
-      );
-      const ok = results.filter((r) => r.status === 'fulfilled').length;
-      const fail = results.length - ok;
-      if (fail === 0) {
-        setToast({
-          open: true,
-          message: `Picks saved to ${ok} group${ok === 1 ? '' : 's'}`,
-          variant: 'success',
-        });
-      } else if (ok === 0) {
-        setToast({
-          open: true,
-          message: `Failed to save picks (0/${results.length})`,
-          variant: 'error',
-        });
-      } else {
-        setToast({
-          open: true,
-          message: `Saved to ${ok}/${results.length} groups (${fail} failed)`,
-          variant: 'error',
-        });
-      }
-    } catch (err) {
-      setToast({
-        open: true,
-        message: err instanceof Error ? err.message : 'Failed to save picks',
-        variant: 'error',
-      });
-    } finally {
-      setSubmitting(false);
-    }
-  }
 
   // No identifier in the query string: nothing to load, show the error UI early.
   if (!identifier) {
@@ -266,92 +42,10 @@ export default function WorldCupPicksPage() {
 
   return (
     <div className="mx-auto max-w-4xl px-sm py-lg sm:p-lg">
-      <h1 className="text-2xl font-bold text-secondary-900 dark:text-neutral-0">
+      <h1 className="mb-lg text-2xl font-bold text-secondary-900 dark:text-neutral-0">
         World Cup 2026 Picks
       </h1>
-
-      {/* Match list, grouped by stage */}
-      <div className="mt-lg space-y-lg">
-        {pageState.loading ? (
-          <p className="py-lg text-center text-secondary">Loading matches…</p>
-        ) : pageState.error ? (
-          <div className="flex flex-col items-center gap-sm py-lg text-center">
-            <p className="text-sm text-error-600 dark:text-error-400">{pageState.error}</p>
-            <Button variant="secondary" size="sm" onClick={() => fetchMatches()}>
-              Try Again
-            </Button>
-          </div>
-        ) : pageState.matches.length === 0 ? (
-          <p className="py-lg text-center text-secondary">No matches found for this tournament.</p>
-        ) : (
-          stageGroups.map((group) => (
-            <section key={group.stage} className="space-y-sm">
-              <h2 className="text-lg font-semibold text-secondary-900 dark:text-neutral-0">
-                {STAGE_LABEL[group.stage]}
-              </h2>
-              {group.matches.map((match) => (
-                <MatchPickRow
-                  key={match.id}
-                  match={match}
-                  pickedResult={draft[match.id] ?? null}
-                  disabled={submitting}
-                  onPick={(result) => pickResult(match.id, result)}
-                />
-              ))}
-            </section>
-          ))
-        )}
-      </div>
-
-      {/* Spacer so the sticky bar never covers the last match row. */}
-      {pageState.matches.length > 0 && <div className="h-20" aria-hidden="true" />}
-
-      {/* Sticky submit bar — pinned to the bottom of the viewport so the
-          user can submit from anywhere in the stage list without scrolling
-          to the page end. */}
-      {pageState.matches.length > 0 && (
-        <div className="sticky bottom-0 -mx-sm sm:-mx-lg mt-lg border-t border-border bg-neutral-0/95 px-sm sm:px-lg py-sm shadow-[0_-2px_8px_-2px_rgba(0,0,0,0.06)] backdrop-blur dark:bg-secondary-900/95">
-          <div className="mx-auto flex max-w-4xl flex-col gap-sm sm:flex-row sm:items-center sm:justify-between">
-            <div className="flex flex-col gap-xxs sm:flex-row sm:items-center sm:gap-md">
-              <span className="text-sm text-secondary">
-                {picks.length} pick{picks.length === 1 ? '' : 's'} selected
-              </span>
-              {groupId && (
-                <SaveTargetsDropdown
-                  groups={myGroups}
-                  sourceIdentifier={groupId}
-                  selected={selectedTargets}
-                  onChange={setSelectedTargets}
-                  label="World Cup"
-                  loading={groupsLoading}
-                  disabled={submitting}
-                />
-              )}
-            </div>
-            <div className="relative inline-toast-anchor">
-              <InlineToast
-                open={toast.open}
-                message={toast.message}
-                variant={toast.variant}
-                onClose={() => setToast((t) => ({ ...t, open: false }))}
-              />
-              <Button
-                variant="primary"
-                size="md"
-                loading={submitting}
-                disabled={!canSubmit}
-                onClick={submit}
-              >
-                {submitting
-                  ? 'Saving…'
-                  : selectedTargets.size > 1
-                    ? `Submit to ${selectedTargets.size}`
-                    : 'Submit Picks'}
-              </Button>
-            </div>
-          </div>
-        </div>
-      )}
+      <WorldCupPicksTab identifier={identifier} />
     </div>
   );
 }
